+++ template_variables/RELEASES.md
@@ -123,3 +123,11 @@ Based on:
 - Speakeasy CLI 1.401.2 (2.421.3) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.7.0] template_variables
+
+## 2024-10-01 00:31:32
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/template-variables.yaml
+- Speakeasy CLI 1.405.6 (2.428.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.7.1] template_variables